Created the CommandInterpreter.java class, renamed the Main class and changed the version to v1.0
