+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>KingdomProgrammers</groupId>
     <artifactId>Medieval-Set-Home</artifactId>
-    <version>v0.6.1</version>
+    <version>v1.0</version>
     <packaging>jar</packaging>
 
     <name>MedievalSetHome</name>
