+++ src/main/java/dansplugins/sethomesystem/HomeRecord.java
@@ -1,4 +1,4 @@
-package sethomesystem;
+package dansplugins.sethomesystem;
 
 import org.bukkit.Location;
 import org.bukkit.World;
